First commit on refactoring the project
