+++ src/main/java/services/gorest/actions/GoRestActions.java
@@ -1,0 +1,7 @@
+package services.gorest.actions;
+
+public abstract class GoRestActions extends GoRestService{
+
+
+
+}
