+++ src/main/java/services/gorest/models/nodes/Avatar.java
@@ -1,4 +1,4 @@
-package services.gorest.pojo.node;
+package services.gorest.models.nodes;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -14,48 +14,22 @@ import org.apache.commons.lang.builder.ToStringBuilder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-        "limit",
-        "remaining",
-        "reset"
+        "href"
 })
-public class RateLimit {
-    @JsonProperty("limit")
-    private Integer limit;
-    @JsonProperty("remaining")
-    private Integer remaining;
-    @JsonProperty("reset")
-    private Integer reset;
+public class Avatar {
+    @JsonProperty("href")
+    private String href;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
-    @JsonProperty("limit")
-    public Integer getLimit() {
-        return limit;
+    @JsonProperty("href")
+    public String getHref() {
+        return href;
     }
 
-    @JsonProperty("limit")
-    public void setLimit(Integer limit) {
-        this.limit = limit;
-    }
-
-    @JsonProperty("remaining")
-    public Integer getRemaining() {
-        return remaining;
-    }
-
-    @JsonProperty("remaining")
-    public void setRemaining(Integer remaining) {
-        this.remaining = remaining;
-    }
-
-    @JsonProperty("reset")
-    public Integer getReset() {
-        return reset;
-    }
-
-    @JsonProperty("reset")
-    public void setReset(Integer reset) {
-        this.reset = reset;
+    @JsonProperty("href")
+    public void setHref(String href) {
+        this.href = href;
     }
 
     @Override
@@ -75,7 +49,7 @@ public class RateLimit {
 
     @Override
     public int hashCode() {
-        return new HashCodeBuilder().append(limit).append(remaining).append(reset).append(additionalProperties).toHashCode();
+        return new HashCodeBuilder().append(href).append(additionalProperties).toHashCode();
     }
 
     @Override
@@ -83,10 +57,10 @@ public class RateLimit {
         if (other == this) {
             return true;
         }
-        if ((other instanceof RateLimit) == false) {
+        if ((other instanceof Avatar) == false) {
             return false;
         }
-        RateLimit rhs = ((RateLimit) other);
-        return new EqualsBuilder().append(limit, rhs.limit).append(remaining, rhs.remaining).append(reset, rhs.reset).append(additionalProperties, rhs.additionalProperties).isEquals();
+        Avatar rhs = ((Avatar) other);
+        return new EqualsBuilder().append(href, rhs.href).append(additionalProperties, rhs.additionalProperties).isEquals();
     }
 }
