+++ src/main/java/services/gorest/models/nodes/Edit.java
@@ -1,4 +1,4 @@
-package services.gorest.pojo.node;
+package services.gorest.models.nodes;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -14,48 +14,22 @@ import org.apache.commons.lang.builder.ToStringBuilder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-        "self",
-        "edit",
-        "avatar"
+        "href"
 })
-public class Links {
-    @JsonProperty("self")
-    private Self self;
-    @JsonProperty("edit")
-    private Edit edit;
-    @JsonProperty("avatar")
-    private Avatar avatar;
+public class Edit {
+    @JsonProperty("href")
+    private String href;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
-    @JsonProperty("self")
-    public Self getSelf() {
-        return self;
+    @JsonProperty("href")
+    public String getHref() {
+        return href;
     }
 
-    @JsonProperty("self")
-    public void setSelf(Self self) {
-        this.self = self;
-    }
-
-    @JsonProperty("edit")
-    public Edit getEdit() {
-        return edit;
-    }
-
-    @JsonProperty("edit")
-    public void setEdit(Edit edit) {
-        this.edit = edit;
-    }
-
-    @JsonProperty("avatar")
-    public Avatar getAvatar() {
-        return avatar;
-    }
-
-    @JsonProperty("avatar")
-    public void setAvatar(Avatar avatar) {
-        this.avatar = avatar;
+    @JsonProperty("href")
+    public void setHref(String href) {
+        this.href = href;
     }
 
     @Override
@@ -75,7 +49,7 @@ public class Links {
 
     @Override
     public int hashCode() {
-        return new HashCodeBuilder().append(self).append(edit).append(avatar).append(additionalProperties).toHashCode();
+        return new HashCodeBuilder().append(href).append(additionalProperties).toHashCode();
     }
 
     @Override
@@ -83,11 +57,11 @@ public class Links {
         if (other == this) {
             return true;
         }
-        if ((other instanceof Links) == false) {
+        if ((other instanceof Edit) == false) {
             return false;
         }
-        Links rhs = ((Links) other);
-        return new EqualsBuilder().append(self, rhs.self).append(edit, rhs.edit).append(avatar, rhs.avatar).append(additionalProperties, rhs.additionalProperties).isEquals();
+        Edit rhs = ((Edit) other);
+        return new EqualsBuilder().append(href, rhs.href).append(additionalProperties, rhs.additionalProperties).isEquals();
     }
