+++ src/main/java/services/gorest/models/nodes/RateLimit.java
@@ -1,4 +1,4 @@
-package services.gorest.pojo.responses;
+package services.gorest.models.nodes;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -11,40 +11,51 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import org.apache.commons.lang.builder.EqualsBuilder;
 import org.apache.commons.lang.builder.HashCodeBuilder;
 import org.apache.commons.lang.builder.ToStringBuilder;
-import services.gorest.pojo.User;
-import services.gorest.pojo.node._Meta;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-        "_meta",
-        "result"
+        "limit",
+        "remaining",
+        "reset"
 })
-public class GetUserResponse {
-    @JsonProperty("_meta")
-    private _Meta _meta;
-    @JsonProperty("result")
-    private User result;
+public class RateLimit {
+    @JsonProperty("limit")
+    private Integer limit;
+    @JsonProperty("remaining")
+    private Integer remaining;
+    @JsonProperty("reset")
+    private Integer reset;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
-    @JsonProperty("_meta")
-    public _Meta get_meta() {
-        return _meta;
+    @JsonProperty("limit")
+    public Integer getLimit() {
+        return limit;
     }
 
-    @JsonProperty("_meta")
-    public void set_meta(_Meta _meta) {
-        this._meta = _meta;
+    @JsonProperty("limit")
+    public void setLimit(Integer limit) {
+        this.limit = limit;
     }
 
-    @JsonProperty("result")
-    public User getResult() {
-        return result;
+    @JsonProperty("remaining")
+    public Integer getRemaining() {
+        return remaining;
     }
 
-    @JsonProperty("result")
-    public void setResult(User result) {
-        this.result = result;
+    @JsonProperty("remaining")
+    public void setRemaining(Integer remaining) {
+        this.remaining = remaining;
+    }
+
+    @JsonProperty("reset")
+    public Integer getReset() {
+        return reset;
+    }
+
+    @JsonProperty("reset")
+    public void setReset(Integer reset) {
+        this.reset = reset;
     }
 
     @Override
@@ -64,7 +75,7 @@ public class GetUserResponse {
 
     @Override
     public int hashCode() {
-        return new HashCodeBuilder().append(_meta).append(result).append(additionalProperties).toHashCode();
+        return new HashCodeBuilder().append(limit).append(remaining).append(reset).append(additionalProperties).toHashCode();
     }
 
     @Override
@@ -72,12 +83,10 @@ public class GetUserResponse {
         if (other == this) {
             return true;
         }
-        if ((other instanceof GetUserResponse) == false) {
+        if ((other instanceof RateLimit) == false) {
             return false;
         }
-        GetUserResponse rhs = ((GetUserResponse) other);
-        return new EqualsBuilder().append(_meta, rhs._meta).append(result, rhs.result).append(additionalProperties, rhs.additionalProperties).isEquals();
+        RateLimit rhs = ((RateLimit) other);
+        return new EqualsBuilder().append(limit, rhs.limit).append(remaining, rhs.remaining).append(reset, rhs.reset).append(additionalProperties, rhs.additionalProperties).isEquals();
     }
-
-
 }
